+++ app/controllers/api/v1/contacts_controller.rb
@@ -1,4 +1,8 @@
 class Api::V1::ContactsController < ApiController
+  def show
+    handle_with_manager Contact::ControllerManager::GetInfo.new(current_user.mkey, params)
+  end
+
   def create
     handle_with_manager Contact::ControllerManager::ValidateRawParams.new(current_user.mkey, params) do
       Resque.enqueue(ResqueWorker::ImportContacts, current_user.mkey, params['contacts'])