+++ app/controllers/api/v1/notifications_controller.rb
@@ -1,20 +1,17 @@
 class Api::V1::NotificationsController < ApiController
   before_action :set_web_client
 
+  def show
+    render json: { status: :success,
+                   data: NotificationSerializer.new(@web_client.notification).serializable_hash }
+  end
+
   def add
     handle_action(:add)
   end
 
   def ignore
     handle_action(:ignore)
-  end
-
-  def unsubscribe
-    handle_action(:unsubscribe)
-  end
-
-  def subscribe
-    handle_action(:subscribe)
   end
 
   private
@@ -25,7 +22,7 @@ class Api::V1::NotificationsController < ApiController
   end
 
   def set_web_client
-    @web_client = WebClient::ActionHandler.new(params[:id])
+    @web_client = WebClient::ActionHandler.new(params[:id], caller: :api)
     render status: :unprocessable_entity, json: { status: :failure, errors: @web_client.errors } unless @web_client.valid?
   end
 end