+++ app/services/web_client/action_handler.rb
@@ -1,16 +1,18 @@
 class WebClient::ActionHandler
   include ActiveModel::Validations
 
-  attr_reader :nkey, :notifications, :notice
+  attr_reader :nkey, :caller, :notifications, :notice
   validate :nkey_should_be_correct
 
-  def initialize(nkey)
+  def initialize(nkey, owner: nil, caller: :web_client)
     @nkey = nkey
+    @owner = owner
+    @caller = caller
     @notifications = Notification.by_nkey(nkey)
   end
 
   def owner
-    notification.contact.owner
+    @owner || notification.contact.owner
   end
 
   def notification
@@ -22,6 +24,7 @@ class WebClient::ActionHandler
   #
 
   def add(contact = nil)
+    emit_event(%w(notification added)) unless contact
     add_or_ignore(Contact::Add, contact, :added)
   rescue AASM::InvalidTransition
     if notification.added?
@@ -31,6 +34,7 @@ class WebClient::ActionHandler
   end
 
   def ignore(contact = nil)
+    emit_event(%w(notification ignored)) unless contact
     add_or_ignore(Contact::Ignore, contact, :ignored)
   rescue AASM::InvalidTransition
     status = notification.status
@@ -41,10 +45,12 @@ class WebClient::ActionHandler
   end
 
   def unsubscribe
+    emit_event(%w(settings unsubscribed))
     subscribe_or_unsubscribe(:unsubscribe, :unsubscribed)
   end
 
   def subscribe
+    emit_event(%w(settings subscribed))
     subscribe_or_unsubscribe(:subscribe, :subscribed)
   end
 
@@ -60,14 +66,14 @@ class WebClient::ActionHandler
                                 contact_name: contact_to_handle.display_name).as_json
 
     update_status(new_status) unless contact
-    service.new(contact_to_handle).do
+    service.new(contact_to_handle, caller: caller).do
   end
 
   def subscribe_or_unsubscribe(action_method, new_status)
-    if notification.contact.owner.send("#{new_status}?")
+    if owner.send("#{new_status}?")
       @notice = NoticeBuilder.new(new_status, new_status, "already_#{new_status}").as_json
     else
-      notification.contact.owner.send(action_method)
+      owner.send(action_method)
       @notice = NoticeBuilder.new(new_status, new_status, new_status).as_json
     end
   end
@@ -85,5 +91,25 @@ class WebClient::ActionHandler
 
   def nkey_should_be_correct
     errors.add(:nkey, 'nkey is incorrect') if notifications.empty?
+  end
+
+  #
+  # events dispatching
+  #
+
+  def emit_event(name)
+    Zazo::Tools::EventDispatcher.emit(name, build_event)
+  end
+
+  def build_event
+    event = {
+      triggered_by: "ff:#{caller}",
+      initiator: 'owner',
+      initiator_id: owner.mkey
+    }
+    event.merge!(
+      target: 'notification',
+      target_id: notification.nkey) if notification
+    event
   end
 end