+++ spec/controllers/api/v1/notifications_controller_spec.rb
@@ -8,12 +8,19 @@ RSpec.describe Api::V1::NotificationsController, type: :controller do
   let(:incorrect_nkey) { 'xxxxxxxxxxxx' }
 
   before do
-    authenticate_with_http_digest(user_mkey, user_auth) { post action, id: nkey }
+    authenticate_with_http_digest(user_mkey, user_auth) { send(*(action + [{ id: nkey }])) }
     notification.reload
   end
 
+  describe 'GET #show' do
+    let(:action) { [:get, :show] }
+
+    it { expect(response).to be_success }
+    it { expect(json_response['data']['contact']).to be_kind_of Hash }
+  end
+
   describe 'POST #add' do
-    let(:action) { :add }
+    let(:action) { [:post, :add] }
 
     it { expect(response).to be_success }
     it { expect(notification.status).to eq 'added' }
@@ -23,29 +30,16 @@ RSpec.describe Api::V1::NotificationsController, type: :controller do
 
       it { expect(response).to be_unprocessable }
       it { expect(notification.status).to eq 'no_feedback' }
-      it { expect(response.body).to eq "{\"status\":\"failure\",\"errors\":{\"nkey\":[\"nkey is incorrect\"]}}" }
+      it { expect(json_response).to eq 'status' => 'failure',
+                                       'errors' => { 'nkey' => ['nkey is incorrect'] } }
       it { expect(response.header['Content-Type']).to include 'application/json'  }
     end
   end
 
   describe 'POST #ignore' do
-    let(:action) { :ignore }
+    let(:action) { [:post, :ignore] }
 
     it { expect(response).to be_success }
     it { expect(notification.status).to eq 'ignored' }
-  end
-
-  describe 'POST #unsubscribe' do
-    let(:action) { :unsubscribe }
-
-    it { expect(response).to be_success }
-    it { expect(notification.status).to eq 'no_feedback' }
-  end
-
-  describe 'POST #subscribe' do
-    let(:action) { :subscribe }
-
-    it { expect(response).to be_success }
-    it { expect(notification.status).to eq 'no_feedback' }
   end
 end